🔨 J'ai mis les fichiers Followers en tant que views pour les mettre en tant que page entière
