+++ src/views/UserFollowers.vue
@@ -1,11 +1,10 @@
 <template>
     <div>
-        <h2>Il/Elle suit: </h2>
+        <h2>Ses followers: </h2>
         <div v-for="post in posts" :key="post.id" >
-            <h2><a v-bind:href="url + post.id">{{ post.firstname}}</a></h2>
-            <div>{{ post.lastname}}</div>
-            <div>{{ post.email}}</div>
             <img v-bind:src="post.picture" v-bind:alt="post.firstname + post.lastname " />
+            <h2><a v-bind:href="url + post.id">{{ post.firstname}} {{ post.lastname}}</a></h2>
+            <hr>
         </div>
     </div>
 </template>
@@ -15,13 +14,13 @@ import { ref, defineComponent } from 'vue';
 import { useRouter } from 'vue-router';
 import api from '../api';
 
-function useUserFollows(userId) {
+function UserFollowers(userId) {
   const posts = ref([]);
   const url = '/profile/';
 
-  api.getRequest(`/user/follow/${userId}`, (data) => {
+  api.getRequest(`/user/followers/${userId}`, (data) => {
     if (data !== 'error') {
-      posts.value = data.follows;
+      posts.value = data.followers;
     }
   });
 
@@ -34,7 +33,7 @@ function useUserFollows(userId) {
 export default defineComponent({
   setup() {
     const { id } = useRouter().currentRoute.value.params;
-    return { ...useUserFollows(id) };
+    return { ...UserFollowers(id) };
   },
 });
 </script>
