+++ src/views/UserFollows.vue
@@ -1,11 +1,9 @@
 <template>
     <div>
-        <h2>Ses followers: </h2>
+        <h2>Il/Elle suit: </h2>
         <div v-for="post in posts" :key="post.id" >
-            <h2><a v-bind:href="url + post.id">{{ post.firstname}}</a></h2>
-            <div>{{ post.lastname}}</div>
-            <div>{{ post.email}}</div>
             <img v-bind:src="post.picture" v-bind:alt="post.firstname + post.lastname " />
+            <h2><a v-bind:href="url + post.id">{{ post.firstname}} {{ post.lastname}}</a></h2>
         </div>
     </div>
 </template>
@@ -15,13 +13,13 @@ import { ref, defineComponent } from 'vue';
 import { useRouter } from 'vue-router';
 import api from '../api';
 
-function UserFollowers(userId) {
+function useUserFollows(userId) {
   const posts = ref([]);
   const url = '/profile/';
 
-  api.getRequest(`/user/followers/${userId}`, (data) => {
+  api.getRequest(`/user/follow/${userId}`, (data) => {
     if (data !== 'error') {
-      posts.value = data.followers;
+      posts.value = data.follows;
     }
   });
 
@@ -34,7 +32,7 @@ function UserFollowers(userId) {
 export default defineComponent({
   setup() {
     const { id } = useRouter().currentRoute.value.params;
-    return { ...UserFollowers(id) };
+    return { ...useUserFollows(id) };
   },
 });
 </script>
